+++ doc/tutorials/GettingStarted/README.md
@@ -71,7 +71,7 @@ Precompute the rotation matrices used internally by COMMIT to create the lookup-
 
 ```python
 import commit
-commit.core.precompute_rotation_matrices()
+commit.lut.precompute_rotation_matrices()
 ```
 
 Now, load the data:
